Added frontend env for source score threshold (#20)
Optional frontend NEXT_PUBLIC_NODE_SCORE_THRESHOLD controls the threshold value which determines when sources are shown in the chat. Lower values mean sources are more likely to be shown and higher values mean sources are more likely not being shown. The new env is optional and default value is 0.25